+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -44,15 +44,19 @@ import kotlin.test.fail
 class HubTest {
 
     private lateinit var file: File
+    private lateinit var profilingTraceFile: File
 
     @BeforeTest
     fun `set up`() {
         file = Files.createTempDirectory("sentry-disk-cache-test").toAbsolutePath().toFile()
+        profilingTraceFile = Files.createTempFile("trace", ".trace").toFile()
+        profilingTraceFile.writeText("sampledProfile")
     }
 
     @AfterTest
     fun shutdown() {
         file.deleteRecursively()
+        profilingTraceFile.delete()
         Sentry.close()
     }
 
@@ -1273,7 +1277,7 @@ class HubTest {
         sentryTracer.finish()
         sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceContext?)
         verify(mockClient, never()).captureTransaction(any(), any(), any())
-        verify(mockClient, never()).captureTransaction(any(), any(), any(), anyOrNull(), anyOrNull())
+        verify(mockClient, never()).captureTransaction(any(), any(), any(), anyOrNull())
     }
 
     @Test
@@ -1289,42 +1293,7 @@ class HubTest {
         val sentryTracer = SentryTracer(TransactionContext("name", "op", TracesSamplingDecision(true)), sut)
         sentryTracer.finish()
         val traceContext = sentryTracer.traceContext()
-        verify(mockClient).captureTransaction(any(), equalTraceContext(traceContext), any(), eq(null), anyOrNull())
-    }
-
-    @Test
-    fun `when startTransaction and profiling is enabled, transaction is profiled only if sampled`() {
-        val mockTransactionProfiler = mock<ITransactionProfiler>()
-        val hub = generateHub {
-            it.setTransactionProfiler(mockTransactionProfiler)
-        }
-        // Transaction is not sampled, so it should not be profiled
-        val contexts = TransactionContext("name", "op", TracesSamplingDecision(false, null, true, null))
-        val transaction = hub.startTransaction(contexts)
-        transaction.finish()
-        verify(mockTransactionProfiler, never()).onTransactionStart(anyOrNull())
-        verify(mockTransactionProfiler, never()).onTransactionFinish(anyOrNull())
-
-        // Transaction is sampled, so it should be profiled
-        val sampledContexts = TransactionContext("name", "op", TracesSamplingDecision(true, null, true, null))
-        val sampledTransaction = hub.startTransaction(sampledContexts)
-        sampledTransaction.finish()
-        verify(mockTransactionProfiler).onTransactionStart(anyOrNull())
-        verify(mockTransactionProfiler).onTransactionFinish(anyOrNull())
-    }
-
-    @Test
-    fun `when startTransaction and is sampled but profiling is disabled, transaction is not profiled`() {
-        val mockTransactionProfiler = mock<ITransactionProfiler>()
-        val hub = generateHub {
-            it.profilesSampleRate = 0.0
-            it.setTransactionProfiler(mockTransactionProfiler)
-        }
-        val contexts = TransactionContext("name", "op")
-        val transaction = hub.startTransaction(contexts)
-        transaction.finish()
-        verify(mockTransactionProfiler, never()).onTransactionStart(anyOrNull())
-        verify(mockTransactionProfiler, never()).onTransactionFinish(anyOrNull())
+        verify(mockClient).captureTransaction(any(), equalTraceContext(traceContext), any(), eq(null))
     }
 
     @Test
@@ -1336,7 +1305,7 @@ class HubTest {
         val sut = Hub(options)
         val mockClient = mock<ISentryClient>()
         sut.bindClient(mockClient)
-        whenever(mockClient.captureTransaction(anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull())).thenReturn(SentryId())
+        whenever(mockClient.captureTransaction(anyOrNull(), anyOrNull(), anyOrNull(), anyOrNull())).thenReturn(SentryId())
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", TracesSamplingDecision(true)), sut)
         sentryTracer.finish()
@@ -1355,7 +1324,7 @@ class HubTest {
 
         val sentryTracer = SentryTracer(TransactionContext("name", "op", TracesSamplingDecision(true)), sut)
         sut.captureTransaction(SentryTransaction(sentryTracer), null as TraceContext?)
-        verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null), anyOrNull())
+        verify(mockClient, never()).captureTransaction(any(), any(), any(), eq(null))
     }
 
     @Test
@@ -1371,7 +1340,7 @@ class HubTest {
         val sentryTracer = SentryTracer(TransactionContext("name", "op", TracesSamplingDecision(false)), sut)
         sentryTracer.finish()
         val traceContext = sentryTracer.traceContext()
-        verify(mockClient, never()).captureTransaction(any(), equalTraceContext(traceContext), any(), eq(null), anyOrNull())
+        verify(mockClient, never()).captureTransaction(any(), equalTraceContext(traceContext), any(), eq(null))
     }
 
     @Test
@@ -1391,6 +1360,63 @@ class HubTest {
             options.clientReportRecorder,
             listOf(DiscardedEvent(DiscardReason.SAMPLE_RATE.reason, DataCategory.Transaction.category, 1))
         )
+    }
+    //endregion
+
+    //region captureProfile tests
+    @Test
+    fun `when captureProfile is called on disabled client, do nothing`() {
+        val options = SentryOptions()
+        options.cacheDirPath = file.absolutePath
+        options.dsn = "https://key@sentry.io/proj"
+        options.setSerializer(mock())
+        val sut = Hub(options)
+        val mockClient = mock<ISentryClient>()
+        sut.bindClient(mockClient)
+        sut.close()
+
+        val sentryTracer = SentryTracer(TransactionContext("name", "op"), sut)
+        sentryTracer.finish()
+        sut.captureProfile(ProfilingTraceData(profilingTraceFile, sentryTracer))
+        verify(mockClient, never()).captureProfile(any())
+    }
+
+    @Test
+    fun `when startTransaction and profiling is enabled, transaction is profiled only if sampled`() {
+        val mockTransactionProfiler = mock<ITransactionProfiler>()
+        whenever(mockTransactionProfiler.onTransactionFinish(any())).thenReturn(mock())
+        val mockClient = mock<ISentryClient>()
+        val hub = generateHub {
+            it.setTransactionProfiler(mockTransactionProfiler)
+        }
+        hub.bindClient(mockClient)
+        // Transaction is not sampled, so it should not be profiled
+        val contexts = TransactionContext("name", "op", TracesSamplingDecision(false, null, true, null))
+        val transaction = hub.startTransaction(contexts)
+        transaction.finish()
+        verify(mockClient, never()).captureProfile(any())
+
+        // Transaction is sampled, so it should be profiled
+        val sampledContexts = TransactionContext("name", "op", TracesSamplingDecision(true, null, true, null))
+        val sampledTransaction = hub.startTransaction(sampledContexts)
+        sampledTransaction.finish()
+        verify(mockClient).captureProfile(any())
+    }
+
+    @Test
+    fun `when startTransaction and is sampled but profiling is disabled, transaction is not profiled`() {
+        val mockTransactionProfiler = mock<ITransactionProfiler>()
+        whenever(mockTransactionProfiler.onTransactionFinish(any())).thenReturn(mock())
+        val mockClient = mock<ISentryClient>()
+        val hub = generateHub {
+            it.profilesSampleRate = 0.0
+            it.setTransactionProfiler(mockTransactionProfiler)
+        }
+        hub.bindClient(mockClient)
+        val contexts = TransactionContext("name", "op")
+        val transaction = hub.startTransaction(contexts)
+        transaction.finish()
+        verify(mockClient, never()).captureProfile(any())
     }
     //endregion